Fix : Remove control character in input data
