+++ src/parsing/parsing.c
@@ -6,7 +6,7 @@
 /*   By: junsan <junsan@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/26 10:49:14 by junsan            #+#    #+#             */
-/*   Updated: 2024/06/11 20:25:11 by junsan           ###   ########.fr       */
+/*   Updated: 2024/06/12 16:02:24 by junsan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
