+++ src/parsing/prints.c
@@ -6,12 +6,25 @@
 /*   By: junsan <junsan@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/26 11:49:26 by junsan            #+#    #+#             */
-/*   Updated: 2024/06/11 12:11:34 by junsan           ###   ########.fr       */
+/*   Updated: 2024/06/12 16:02:22 by junsan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
+static int	is_control_character(char *str)
+{
+	size_t	i;
+
+	i = 0;
+	while (str[i])
+	{
+		if (str[i] < 32 || str[i] == 127)
+			return (0);
+		i++;
+	}
+	return (1);
+}
 void	print_token(t_token *head)
 {
 	if (!head)
@@ -21,6 +34,7 @@ void	print_token(t_token *head)
 	}
 	while (head)
 	{
+		printf("check : %d\n", is_control_character(head->data));
 		printf("data : %s\n", head->data);
 		head = head->next;
 	}
