+++ src/parsing/tokenize.c
@@ -6,7 +6,7 @@
 /*   By: junsan <junsan@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/25 15:39:22 by junsan            #+#    #+#             */
-/*   Updated: 2024/06/11 19:43:55 by junsan           ###   ########.fr       */
+/*   Updated: 2024/06/12 16:02:15 by junsan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
