+++ src/utils/string_utils.c
@@ -6,36 +6,11 @@
 /*   By: junsan <junsan@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/26 15:47:35 by junsan            #+#    #+#             */
-/*   Updated: 2024/06/11 19:47:57 by junsan           ###   ########.fr       */
+/*   Updated: 2024/06/12 16:02:20 by junsan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-char	*trim_whitespace(const char *str)
-{
-	const char	*end;
-
-	while (ft_isspace((unsigned char)*str))
-		str++;
-	end = str + ft_strlen(str) - 1;
-	while (end > str && ft_isspace((unsigned char)*end))
-		end--;
-	return (ft_substr(str, 0, end - str + 1));
-}
-
-char	*trim_first_last(char *str)
-{
-	size_t	len;
-	char	*new_str;
-
-	len = ft_strlen(str);
-	new_str = (char *)malloc(sizeof(char) * (len - 1));
-	if (new_str == NULL)
-		return (NULL);
-	ft_strlcpy(new_str, str + 1, len - 1);
-	return (new_str);
-}
 
 bool	ft_isspace(char c)
 {
@@ -54,4 +29,25 @@ bool	is_all_whitespace(const char *str)
 		str++;
 	}
 	return (true);
+}
+
+static int	is_control_character(char c)
+{
+	return (c < 32 || c == 127);
+}
+
+void	remove_control_characters(char *str)
+{
+	char *src;
+	char *dst;
+
+	src = str;
+	dst = str;
+	while (*src)
+	{
+		if (!is_control_character(*src))
+			*dst++ = *src;
+		src++;
+	}
+	*dst = '\0';
 }
