+++ src/utils/string_utils_2.c
@@ -1,0 +1,38 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   string_utils_2.c                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: junsan <junsan@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/06/12 15:29:43 by junsan            #+#    #+#             */
+/*   Updated: 2024/06/12 15:31:32 by junsan           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+char	*trim_whitespace(const char *str)
+{
+	const char	*end;
+
+	while (ft_isspace((unsigned char)*str))
+		str++;
+	end = str + ft_strlen(str) - 1;
+	while (end > str && ft_isspace((unsigned char)*end))
+		end--;
+	return (ft_substr(str, 0, end - str + 1));
+}
+
+char	*trim_first_last(char *str)
+{
+	size_t	len;
+	char	*new_str;
+
+	len = ft_strlen(str);
+	new_str = (char *)malloc(sizeof(char) * (len - 1));
+	if (new_str == NULL)
+		return (NULL);
+	ft_strlcpy(new_str, str + 1, len - 1);
+	return (new_str);
+}
